main - webgl now works as a "module"
This is dumb - 0 experience points. This is just an unnecessary step since all I do is
export everything that has a dependency, and then import it. The only way I can see to
make this worthwhile from an API standpoint is... not a clear path at all. Everything
requires that context object, and it might just get lost anyway. The API is so completely
simplified by keeping "context" a relative global, it's not worth trying to navigate
passing it as a parameter into everything.